requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
duckdb==1.4.0
et_xmlfile==2.0.0
idna==3.10
ijson==3.4.0
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.2
polars==1.33.1
psycopg2-binary==2.9.10
pyarrow==21.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
